making fast plotting default
