+++ python/plotter/HistCreator.py
@@ -1,4 +1,6 @@
 import hashlib
+from multiprocessing import Pool, Process
+#from multiprocessing.dummy import Pool
 
 from array import array
 
@@ -13,7 +15,6 @@ from pdb import set_trace
 
 from ROOT import TH1F, TFile, TTree, TTreeFormula
 
-
 def initHist(hist, vcfg):
     hist.Sumw2()
     xtitle = vcfg.xtitle
@@ -21,139 +22,6 @@ def initHist(hist, vcfg):
         xtitle += ' ({})'.format(vcfg.unit)
     hist.GetXaxis().SetTitle(xtitle)
     hist.SetStats(False)
-
-
-def createHistograms(hist_cfg, all_stack=False, verbose=False, friend_func=None, vcfgs=None):
-    '''Method to create actual histogram (DataMCPlot) instances from histogram 
-    config; this version handles multiple variables via MultiDraw.
-    '''
-    if hist_cfg.vars:
-        vcfgs = hist_cfg.vars
-
-    if not vcfgs:
-        print 'ERROR in createHistograms: No variable configs passed', hist_cfg.name
-
-    plots = {}
-
-    for vcfg in vcfgs:
-        plot = DataMCPlot(vcfg.name)
-        plot.lumi = hist_cfg.lumi
-        if vcfg.name in plots:
-            print 'Adding variable with same name twice', vcfg.name, 'not yet foreseen; taking the last'
-        plots[vcfg.name] = plot
-
-    for cfg in hist_cfg.cfgs:
-        # First check whether it's a sub-histo or not
-        if isinstance(cfg, HistogramCfg):
-            hists = createHistograms(cfg, all_stack=True, vcfgs=vcfgs)
-            for vcfg in vcfgs:
-                hist = hists[vcfg.name]
-                plot = plots[vcfg.name]
-                hist._BuildStack(hist._SortedHistograms(), ytitle='Events')
-
-                total_hist = plot.AddHistogram(cfg.name, hist.stack.totalHist.weighted, stack=True)
-
-                if cfg.norm_cfg is not None:
-                    norm_hist = createHistogram(cfg.norm_cfg, all_stack=True)
-                    norm_hist._BuildStack(norm_hist._SortedHistograms(), ytitle='Events')
-                    total_hist.Scale(hist.stack.integral/total_hist.Yield())
-
-                if cfg.total_scale is not None:
-                    total_hist.Scale(cfg.total_scale)
-                    # print 'Scaling total', hist_cfg.name, 'by', cfg.total_scale
-        else:
-            print('building histo %s'%cfg.name)
-            # It's a sample cfg
-
-            # Now read the tree
-            file_name = '/'.join([cfg.ana_dir, cfg.dir_name, cfg.tree_prod_name, 'tree.root'])
-
-            # attach the trees to the first DataMCPlot
-            plot = plots[vcfgs[0].name]
-            ttree = plot.readTree(file_name, cfg.tree_name, verbose=verbose, friend_func=friend_func)
-
-            norm_cut = hist_cfg.cut
-            shape_cut = hist_cfg.cut
-
-            if cfg.norm_cut:
-                norm_cut = cfg.norm_cut
-
-            if cfg.shape_cut:
-                shape_cut = cfg.shape_cut
-
-            weight = hist_cfg.weight
-            if cfg.weight_expr:
-                weight = '*'.join([weight, cfg.weight_expr])
-
-            if cfg.cut_replace_func:
-                norm_cut = cfg.cut_replace_func(norm_cut)
-                shape_cut = cfg.cut_replace_func(norm_cut)
-
-            if hist_cfg.weight:
-                norm_cut = '({c}) * {we}'.format(c=norm_cut, we=weight)
-                shape_cut = '({c}) * {we}'.format(c=shape_cut, we=weight)
-
-            # print '#### FULL CUT ####', norm_cut
-
-            # Initialise all hists before the multidraw
-            hists = {}
-
-            for vcfg in vcfgs:
-                # plot = plots[vcfg.name]
-
-                hname = '_'.join([hist_cfg.name, hashlib.md5(hist_cfg.cut).hexdigest(), cfg.name, vcfg.name, cfg.dir_name])
-                if any(str(b) == 'xmin' for b in vcfg.binning):
-                    hist = TH1F(hname, '', vcfg.binning['nbinsx'],
-                                vcfg.binning['xmin'], vcfg.binning['xmax'])
-                else:
-                    hist = TH1F(hname, '', len(vcfg.binning)-1, vcfg.binning)
-
-                initHist(hist, vcfg)
-                hists[vcfg.name] = hist
-
-            var_hist_tuples = []
-
-            for vcfg in vcfgs:
-                # var_hist_tuples.append(('{var} >> {hist}'.format(var=vcfg.drawname, hist=hists[vcfg.name].GetName()), '1.'))
-                var_hist_tuples.append('{var} >> {hist}'.format(var=vcfg.drawname, hist=hists[vcfg.name].GetName()))
-
-            # Implement the multidraw.
-            ttree.MultiDraw(var_hist_tuples, norm_cut)
-
-            # Do another multidraw here, if needed, and reset the scales in a separate loop
-            if shape_cut != norm_cut:
-                scale = hist.Integral()
-                ttree.Project(hname, vcfg.drawname, shape_cut)
-                hist.Scale(scale/hist.Integral())
-
-            stack = all_stack or (not cfg.is_data and not cfg.is_signal)
-
-            # Loop again over the variables and add histograms to plots one by one
-            for vcfg in vcfgs:
-                hist = hists[vcfg.name]
-                plot = plots[vcfg.name]
-
-                hist.Scale(cfg.scale)
-
-                if cfg.name in plot:
-                    print 'Histogram', cfg.name, 'already exists; adding...', cfg.dir_name
-                    hist_to_add = Histogram(cfg.name, hist)
-                    if not cfg.is_data:
-                        hist_to_add.SetWeight(hist_cfg.lumi*cfg.xsec/cfg.sumweights)
-                    plot[cfg.name].Add(hist_to_add)
-                else:
-                    plot_hist = plot.AddHistogram(cfg.name, hist, stack=stack)
-                    print('added histo %s'%vcfg.name)
-
-                    if not cfg.is_data:
-                        plot_hist.SetWeight(hist_cfg.lumi*cfg.xsec/cfg.sumweights)
-
-    set_trace()
-    for plot in plots.itervalues():
-        plot._ApplyPrefs()
-        plot.Draw()
-    return plots
-
 
 def createHistogram(hist_cfg, all_stack=False, verbose=False, friend_func=None):
     '''Method to create actual histogram (DataMCPlot) instance from histogram 
@@ -178,7 +46,6 @@ def createHistogram(hist_cfg, all_stack=False, verbose=False, friend_func=None):
             if cfg.total_scale is not None:
                 total_hist.Scale(cfg.total_scale)
         else:
-            print('building histo %s'%cfg.name)
             # It's a sample cfg
             hname = '_'.join([hist_cfg.name, hashlib.md5(hist_cfg.cut).hexdigest(), cfg.name, vcfg.name, cfg.dir_name])
             if any(str(b) == 'xmin' for b in vcfg.binning):
@@ -225,7 +92,6 @@ def createHistogram(hist_cfg, all_stack=False, verbose=False, friend_func=None):
 
             hist.Scale(cfg.scale)
 
-
             if cfg.name in plot:
                 print 'Histogram', cfg.name, 'already exists; adding...', cfg.dir_name
                 hist_to_add = Histogram(cfg.name, hist)
@@ -234,12 +100,10 @@ def createHistogram(hist_cfg, all_stack=False, verbose=False, friend_func=None):
                 plot[cfg.name].Add(hist_to_add)
             else:
                 plot_hist = plot.AddHistogram(cfg.name, hist, stack=stack)
-                print('added histo %s'%vcfg.name)
 
                 if not cfg.is_data:
                     plot_hist.SetWeight(hist_cfg.lumi*cfg.xsec/cfg.sumweights)
 
-    set_trace()
     plot._ApplyPrefs()
     return plot
 
@@ -347,4 +211,200 @@ def createTrees(hist_cfg, out_dir, verbose=False, friend_func=None):
         out_file.Write()
         out_file.Close()
     return plot
+
+class CreateHists(object):
+    def __init__(self, hist_cfg):
+        self.hist_cfg = hist_cfg
+        if self.hist_cfg.vars:
+            self.vcfgs = hist_cfg.vars
+
+        if not self.vcfgs:
+            print 'ERROR in createHistograms: No variable configs passed', self.hist_cfg.name
+
+        self.plots = {}
+
+        for vcfg in self.vcfgs:
+            plot = DataMCPlot(vcfg.name)
+            plot.lumi = hist_cfg.lumi
+            if vcfg.name in self.plots:
+                print 'Adding variable with same name twice', vcfg.name, 'not yet foreseen; taking the last'
+            self.plots[vcfg.name] = plot
+
+    def createHistograms(self, hist_cfg, all_stack=False, verbose=False, friend_func=None, vcfgs=None):
+        '''Method to create actual histogram (DataMCPlot) instances from histogram 
+        config; this version handles multiple variables via MultiDraw.
+        '''
+        
+#        set_trace()
+        pool = Pool(processes=len(self.hist_cfg.cfgs))
+        print('number of processes for filling histos (ie. samples): %i'%len(self.hist_cfg.cfgs))
+
+        result = pool.map(self.makealltheplots, self.hist_cfg.cfgs) # for no return 
+#        result = pool.apply_async(self.makealltheplots, self.hist_cfg.cfgs) # for no return 
+#        self.plots = result.get() 
+#        set_trace()
+        for i, cfg in enumerate(self.hist_cfg.cfgs):
+            stack = not cfg.is_data and not cfg.is_signal
+#            print(cfg.name, stack)
+            for vcfg in self.vcfgs:
+                hist = result[i][vcfg.name].histos[0].obj # result[0]['CR_hnl_m_12'].histos[0]
+#                hist = hists[vcfg.name]
+                plot = self.plots[vcfg.name]
+
+#                hist.Scale(cfg.scale)
+#                set_trace()
+                if cfg.name in plot:
+                    print 'Histogram', cfg.name, 'already exists; adding...', cfg.dir_name
+                    hist_to_add = Histogram(cfg.name, hist)
+                    if not cfg.is_data:
+                        hist_to_add.SetWeight(hist_cfg.lumi*cfg.xsec/cfg.sumweights)
+                    plot[cfg.name].Add(hist_to_add)
+                else:
+#                    print(cfg.name, hist.GetEntries(), stack)
+                    plot_hist = plot.AddHistogram(cfg.name, hist, stack=stack)
+#                    print('added histo %s'%vcfg.name)
+
+                    if not cfg.is_data:
+                        plot_hist.SetWeight(self.hist_cfg.lumi*cfg.xsec/cfg.sumweights)
+#                print(cfg.name, vcfg.name, len(plot.histos))
+        print('initializing histos done, making stacks...')
+
+        for plot in self.plots.itervalues():
+            plot._ApplyPrefs()
+
+        print('number of processes for drawing (ie. stacks to draw): %i'%len(self.plots))
+        procs = []
+        for i, plot in enumerate(self.plots.itervalues()):
+            proc = Process(target=plot.Draw, args=())
+            procs.append(proc)
+            proc.start()
+     
+        for proc in procs:
+            proc.join()       
+
+#        for plot in self.plots.itervalues():
+#            plot._ApplyPrefs()
+#            plot.Draw()
+        return self.plots
+
+    def makealltheplots(self, cfg):
+        verbose=False
+        friend_func=None
+        all_stack=False
+    #    for cfg in [hist_cfg.cfgs[0]]:
+    #    for cfg in hist_cfg.cfgs:
+            # First check whether it's a sub-histo or not
+        if isinstance(cfg, HistogramCfg):
+            hists = createHistograms(cfg, all_stack=True, vcfgs=self.vcfgs)
+            for h in hists: print(h)
+            for vcfg in self.vcfgs:
+                hist = hists[vcfg.name]
+                plot = self.plots[vcfg.name]
+                set_trace()
+                hist._BuildStack(hist._SortedHistograms(), ytitle='Events')
+                print('stack built')
+                total_hist = plot.AddHistogram(cfg.name, hist.stack.totalHist.weighted, stack=True)
+
+                if cfg.norm_cfg is not None:
+                    norm_hist = createHistogram(cfg.norm_cfg, all_stack=True)
+                    norm_hist._BuildStack(norm_hist._SortedHistograms(), ytitle='Events')
+                    total_hist.Scale(hist.stack.integral/total_hist.Yield())
+
+                if cfg.total_scale is not None:
+                    total_hist.Scale(cfg.total_scale)
+                    # print 'Scaling total', hist_cfg.name, 'by', cfg.total_scale
+        else:
+            print('building histos for %s'%cfg.name)
+            # It's a sample cfg
+
+            # Now read the tree
+            file_name = '/'.join([cfg.ana_dir, cfg.dir_name, cfg.tree_prod_name, 'tree.root'])
+
+            # attach the trees to the first DataMCPlot
+            plot = self.plots[self.vcfgs[0].name]
+#            set_trace()
+            ttree = plot.readTree(file_name, cfg.tree_name, verbose=verbose, friend_func=friend_func)
+
+            norm_cut = self.hist_cfg.cut
+            shape_cut = self.hist_cfg.cut
+
+            if cfg.norm_cut:
+                norm_cut = cfg.norm_cut
+
+            if cfg.shape_cut:
+                shape_cut = cfg.shape_cut
+
+            weight = self.hist_cfg.weight
+            if cfg.weight_expr:
+                weight = '*'.join([weight, cfg.weight_expr])
+
+            if cfg.cut_replace_func:
+                norm_cut = cfg.cut_replace_func(norm_cut)
+                shape_cut = cfg.cut_replace_func(norm_cut)
+
+            if self.hist_cfg.weight:
+                norm_cut = '({c}) * {we}'.format(c=norm_cut, we=weight)
+                shape_cut = '({c}) * {we}'.format(c=shape_cut, we=weight)
+
+            # print '#### FULL CUT ####', norm_cut
+
+            # Initialise all hists before the multidraw
+            hists = {}
+
+            for vcfg in self.vcfgs:
+                # plot = self.plots[vcfg.name]
+
+                hname = '_'.join([self.hist_cfg.name, hashlib.md5(self.hist_cfg.cut).hexdigest(), cfg.name, vcfg.name, cfg.dir_name])
+                if any(str(b) == 'xmin' for b in vcfg.binning):
+                    hist = TH1F(hname, '', vcfg.binning['nbinsx'],
+                                vcfg.binning['xmin'], vcfg.binning['xmax'])
+                else:
+                    hist = TH1F(hname, '', len(vcfg.binning)-1, vcfg.binning)
+
+                initHist(hist, vcfg)
+                hists[vcfg.name] = hist
+
+            var_hist_tuples = []
+
+            for vcfg in self.vcfgs:
+                # var_hist_tuples.append(('{var} >> {hist}'.format(var=vcfg.drawname, hist=hists[vcfg.name].GetName()), '1.'))
+    #                pool.map(var_hist_tuples.append,('{var} >> {hist}'.format(var=vcfg.drawname, hist=hists[vcfg.name].GetName())))
+                var_hist_tuples.append('{var} >> {hist}'.format(var=vcfg.drawname, hist=hists[vcfg.name].GetName()))
+
+            # Implement the multidraw.
+            ttree.MultiDraw(var_hist_tuples, norm_cut)
+
+            # Do another multidraw here, if needed, and reset the scales in a separate loop
+            if shape_cut != norm_cut:
+                scale = hist.Integral()
+                ttree.Project(hname, vcfg.drawname, shape_cut)
+                hist.Scale(scale/hist.Integral())
+
+            stack = all_stack or (not cfg.is_data and not cfg.is_signal)
+
+            # Loop again over the variables and add histograms to self.plots one by one
+            for vcfg in self.vcfgs:
+                hist = hists[vcfg.name]
+                plot = self.plots[vcfg.name]
+
+                hist.Scale(cfg.scale)
+
+                if cfg.name in plot:
+                    print 'Histogram', cfg.name, 'already exists; adding...', cfg.dir_name
+                    hist_to_add = Histogram(cfg.name, hist)
+                    if not cfg.is_data:
+                        hist_to_add.SetWeight(hist_cfg.lumi*cfg.xsec/cfg.sumweights)
+                    plot[cfg.name].Add(hist_to_add)
+                else:
+#                    print(cfg.name, hist.GetEntries(), stack)
+                    plot_hist = plot.AddHistogram(cfg.name, hist, stack=stack)
+#                    print('added histo %s for %s'%(vcfg.name,cfg.name))
+
+                    if not cfg.is_data:
+                        plot_hist.SetWeight(self.hist_cfg.lumi*cfg.xsec/cfg.sumweights)
+            print('added histos for %s'%cfg.name)
+            PLOTS = self.plots
+        return PLOTS
+
+
